separate components and create cropper and modal components and optimization clean code
