+++ package.json
@@ -43,6 +43,7 @@
     ]
   },
   "devDependencies": {
-    "@types/react-modal": "^3.13.1"
+    "@types/react-modal": "^3.13.1",
+    "@types/styled-components": "^5.1.26"
   }
 }
