+++ src/components/Cropper/index.ts
@@ -1,0 +1,5 @@
+import Cropper from "../Modals/modals/CropperModal";
+
+export {
+    Cropper
+}
