+++ src/components/Modals/index.ts
@@ -1,0 +1,6 @@
+import Modal from "./Modal";
+export * from './modals';
+
+export {
+    Modal
+}
