+++ src/components/Modals/modals/index.ts
@@ -1,0 +1,5 @@
+import CropperModal from "./CropperModal";
+
+export {
+    CropperModal
+}
