--- src/services/useDebounceEffect.ts
@@ -1,17 +1,0 @@
-import { useEffect, DependencyList } from 'react'
-
-export function useDebounceEffect(
-  fn: () => void,
-  waitTime: number,
-  deps?: any,
-) {
-  useEffect(() => {
-    const t = setTimeout(() => {
-      fn.apply(undefined, deps)
-    }, waitTime)
-
-    return () => {
-      clearTimeout(t)
-    }
-  }, deps)
-}
